requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 爬虫相关
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
html5lib==1.1

# AI处理
openai==1.3.5
tiktoken==0.5.2

# 定时任务
apscheduler==3.10.4

# 工具库
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1

# 日志和监控
loguru==0.7.2

# 文本处理
jieba==0.42.1  # 中文分词（可选）
nltk==3.8.1    # 英文文本处理（可选）

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

